Use more C variables in SART inner loops
There were a couple of instances in the SART inner loops where Python
variables were used, either because the variable hadn't been declared
as a C variable or because the Python version rather than the C version
of the variable was used.

Using the C variables instead results in an almost 5-fold speed increase
on a test of inverting 32 LOS measurements on 900 cells, with no discernable
change in output.