Enable saving bolometer cameras in compressed pickle format
Bolometer sensitivity matrices are typically sparse, and so
compress very well. Compressing the pickle file a camera is
saved in can therefore result in significant space savings.
The file extension proposed is .pgz, analogous to .tgz for
tarballs also compressed with gzip.

This is particularly relevant in the following use case: split
a sensitivity calculation on a large grid into many separate jobs,
each of which calculates only a subset of the grid. Write an output
pickle file for each job, in which the sensitivity matrix is all zeros
and only the indices of the grid cells involved in the job are
updated. In post processing, sum the matrices for all the separate
jobs to get the total sensitivity for the grid. In this case, having
many uncompressed pickle files is extremely space inefficient, and
can even lead to storage quota issues if enough jobs are run.
Compressing the intermediate pickle files mitigates this issue.