Enable specifying the render engine to use for bolometer observations
In many use cases, bolometer calculations involve making a large number
of observations, where each observation involves relatively few rays.
In these cases, it is significantly more efficient to parallelise
over observations, rather the parallelising the observations
themselves.

This commit enables specifying the render engine to use for the
LOS and volume observers associated with the bolometer detectors.
Code which instantiates a bolometer camera then then choose to use
a parallel render engine (the default), or use a serial render engine
and parallelise the job at a higher level.