Add option to calculate bolometer sensitivities for a subset of grid cells
For particularly large grids, significant parallel speedups can be made
by breaking the grid into separate chunks and calculating the sensitivities
of the cells in these chunks in parallel jobs. This commit modifies the
calculate_sensitivity method to allow scripts which require this feature
to specify which subset of the grid a particular job should calculate.

It is left to the calling software to handle splitting the calculation
and combining the result.