Add support for describing Axisymmetric voxels using CSG
Using CSG primitives to describe Axisymmetric voxels results in
significantly lower memory usage than using meshes. The difference is
particularly pronounced for voxels with a radial width significantly
less than the mean radius of the voxel. For example, for a grid of
voxels of major radius ~1m on average:
- voxel width 5cm, CSG uses 6% the memory of mesh
- voxel width 2cm, CSG uses 2% the memory of mesh
- voxel width 2mm, 750k cell grid uses ~4.5GB of memory (mesh grid doesn't
  fit into memory on 128GB node, but a ~5000-cell subset uses ~8GB)

The reduction in memory usage makes it easier to work with high-voxel-count
grids, since they do not need to be split up. The CSG forumlation reduces the
child count of the voxels, which in turn speeds up pickling them and reduces
the file size of pickle save files.

Voxels with arbitrary polygon-shaped cross sections are supported, by splitting
the cross section into triangles. Each sub-triangle is used to form a toroidally-
symmetric sub-voxel using CSG operations on cylinders and cones, and these
sub-voxels are made children of the parent voxel.

The use of cone primitives does impact performance, as the KD tree is less
able to optimize ray tracing operations if the height of the cones are larger
than the resulting CSG voxel. As a performance enhancement, rectangular voxels
with edges on the x and z axes are given special treatment: instead of being
split into triangles, these are implemented as the subtraction of 2 cylinders.
This means that for voxel grids composed mainly of rectangular cells, with
perhaps non-rectangular or non-aligned cells at the edges of the grid, performance
is similar to the mesh formulation (but with significantly reduced memory overhead).

The choice of formulation (mesh or CSG) is chosen when the voxel is instantiated,
and defaults to mesh to maintain previous behaviour.

Also included are a series of tests. The principle of the tests is simple:
generate a series of points (r_i, 0, z_i), and check that the set of these
points which lie inside the voxel matches the set of points (r_i, z_i) which
lie in a 2D polygon of the same cross section. All possible triangle shapes
are tested, as well as a few extra cases of rectangular and pentagonal shapes.