Make the least squares inversion API more consistent
Make the API match that of the regularised SART inversion: the scale
factor controlling the relative importance of the measurement and
regularisation is always applied to the Tikhonov matrix, whether it is
specified or not. If unspecified, the Tikhonov matrix defaults to the
identity matrix.

This means callers need to provide an un-scaled Tikhonov matrix if
they want a custom regularisation scheme (such as isotropic smoothness
or ADMT).