Add calculation of cross section centroid for voxels
This gives the "centre of mass" of the voxel in the R, Z plane, using
a formula by Bourke which is accurate for a voxel of arbitrary polygon
cross section shape. Unit tests for this and the area calculation have
been added, as well as some typing for small performance improvements
in the algorithms.