Improve calculation of voxel emissivities from function
Move calculation for each voxel into the Voxel class, and provide an
implementation for the AxisymmetricVoxel. Include typing to improve
performance, and also include unit tests. Fix Github issue #142
ensuring all sub-triangles in the cross section are sampled
correctly. Update documentation for the AxissymmetricVoxel
implementation to clarify the required symmetry of the callable
function.