Enable running emissivity tests without quadpy
The quadpy package is used to numerically integrate the functions in
the emissivity tests to compare with Cherab's Monte-Carlo sampling
technique. Since quadpy is not a Cherab dependency, the results of the
integration are stored in look-up tables for when the quadpy package
is not installed. However, the full calculation is also kept for
provenance reasons, and will be run if quadpy is available.