Change: material and references
Laser material simplified. Most of the calculations moved to the scattering model. Caches of plasma and laser transforms
ScatteringModel got reference to the plasma and laser_model to provide values for the scattering calculations
ScatteringModel changed to LaserEmissionModel
Model manager handling emission models added to laser
Propagation of changes and notifications changed to be more like the beam model.