Make IRVB geometry be specified by an affine transform
This maintains consistency with the rest of the Raysect/Cherab
framework, at the cost of breaking consistency with BolometerCamera
and BolometerFoil. However, the intention is to change the rest of the
bolometer observers to requre a transform rather than centre and basis
vectors at a later date, so implementing this new API for the IRVB
from the start will avoid a breaking change later.

The docstring of the IRVB includes an example of how to generate the
affine transform from a centre point and two basis vectors, so it
should not be difficult for users already familiar with bolometry in
Cherab to adapt to this change.