Feature/IRVB (#206)
* Add IRVB bolometer detector

* Make IRVB and Foil bolometer constructors mostly identical.

Clean up some un-needed instance attributes in both, and ensure that
the IRVB constructor does essentially the same thing as the foil
bolometer: they should both work almost the same way, after all.

* Make pixels_as_foils return an array; tidy up documentation.

pixels_as_foils now returns a Numpy array of objects, for more
powerful indexing.

* Fix setting units and accumulate in BolometerIRVB.

* Pass pixel index rather than pixel to worker for IRVB etendue.

This means the pixel BolometerFoil instance does not need to be pickled:
it will be available in the child process's memory. This avoids errors
related to missing __reduce_cython__ methods.

* Tidy up IRVB docs, add a demo script.

* Fix bug in IRVB demo where port was not cutting through wall.

Remove outer wall and just cut port out of first wall. The
axisymmetric_mesh_from_polygon function already creates a hollow mesh,
so the subtraction will cut through the wall correctly.

* Address comments from PR review

- Switch from strings to an Enum for IRVB units options. Make
  specifying units by end users case-insensitive.
- Replace dectector_id with name, for consistency with other
  observers.
- Replace type checking in `__init__` method with attempted
  conversion to the correct type. Let the conversion functions catch
  incompatible types.
- Reduce targetted_path_prob to 0.99: the value of 1.0 is not robust
  to all slit size/distance combinations and correct instantiation of
  the camera using modern recommended techniques does not require 1.0.
- Remove rows and columns arguments from trace_sightlines(): the bug
  which caused poor performance in BolometerFoil.trace_sightlines has
  now been fixed so sightlines can be indexed after all have been
  traced, with negligible performance penalty.
- Use None for default arguments in calculate_sensitivity and
  calculate_etendue, and set the actual defaults in the method body.

* Fix IRVB demo compatibility with raysect 0.7

* Fix setting of IRVB accumulate property

* Make IRVB geometry be specified by an affine transform

This maintains consistency with the rest of the Raysect/Cherab
framework, at the cost of breaking consistency with BolometerCamera
and BolometerFoil. However, the intention is to change the rest of the
bolometer observers to requre a transform rather than centre and basis
vectors at a later date, so implementing this new API for the IRVB
from the start will avoid a breaking change later.

The docstring of the IRVB includes an example of how to generate the
affine transform from a centre point and two basis vectors, so it
should not be difficult for users already familiar with bolometry in
Cherab to adapt to this change.