Remove Cython as a runtime dependency
Cython is now only a build-time dependency, as specified by PEP518. It
is not required for running Cherab, so has been removed from
install_requires.

A quick check showed there was one module, radiation_function.pyx, which
imported Cython at runtime. That was unnecessary, so it's been changed
to cimport Cython at build time instead. As an added bonus, a small
performance improvement (using a Cython accessor function rather than a
Python property) has been made, and a pxd file to enable the module to
be cimported has been added.