Fix update bug of LaserMaterial
Laser primitives have to be rebuild with the materials.
Possible source of the problem:
Laser node is the parent of the laser primitives
After laser.transform is changed the _modified method is called before
the transforms of the laser primitives is changed because the Laser node
is above the laser primitives in the scenegraph.